Create Meeting Room Regulation
Write a program that determines whether a meeting room is in violation of the fire law regulations regarding the maximum room capacity. The program will read in the maximum room capacity and the number of people to attend the meeting.
If the number of people is less than or equal to the maximum room capacity, the program announces that it is legal to hold the meeting and tells how many additional people may legally attend.
If the number exceeds the maximum room capacity, the program announces that the meeting cannot be held as planned due to fire relations and tell how many people must be excluded in order to meet the fire regulations.